app/requirements.txt
altgraph==0.17.3
appnope==0.1.3
asttokens==2.0.5
backcall==0.2.0
certifi==2022.6.15
debugpy==1.6.2
decorator==5.1.1
distlib==0.3.5
entrypoints==0.4
executing==0.8.3
filelock==3.7.1
ipykernel==6.15.1
ipython==8.4.0
jedi==0.18.1
jupyter-client==7.3.4
jupyter-core==4.11.1
macholib==1.16.2
matplotlib-inline==0.1.3
modulegraph==0.19.3
nest-asyncio==1.5.5
numpy==1.24.1
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pipenv==2022.7.24
platformdirs==2.5.2
prompt-toolkit==3.0.30
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py2app==0.28.5
pyglet==2.0.3
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pyzmq==23.2.0
six==1.16.0
stack-data==0.3.0
tornado==6.2
traitlets==5.3.0
virtualenv==20.16.2
virtualenv-clone==0.5.7
wcwidth==0.2.5
